backend/requirements.txt
# SmartTask Pro — Backend
fastapi==0.115.0
uvicorn[standard]==0.30.0
psycopg2-binary>=2.9.9
psycopg[binary]>=3.2.10
python-dotenv==1.0.1
pydantic==2.9.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
gunicorn>=21.2.0
# Google Calendar Integration
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-httplib2>=0.1.0
